+++ DependencyInjection/CompilerPass/FileLocatorCompilerPass.php
@@ -2,35 +2,24 @@
 
 namespace Doppy\PdfGeneratorBundle\DependencyInjection\CompilerPass;
 
-use Doppy\UtilBundle\Helper\CompilerPass\BaseTagServiceCompilerPass;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\Compiler\PriorityTaggedServiceTrait;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Definition;
-use Symfony\Component\DependencyInjection\Reference;
 
-class FileLocatorCompilerPass extends BaseTagServiceCompilerPass
+class FileLocatorCompilerPass implements CompilerPassInterface
 {
-    protected function handleTag(
-        ContainerBuilder $containerBuilder,
-        Definition $serviceDefinition,
-        Reference $taggedServiceReference,
-        $attributes
-    )
-    {
-        $serviceDefinition->addMethodCall(
-            'addLocator',
-            array(
-                $taggedServiceReference
-            )
-        );
-    }
+    use PriorityTaggedServiceTrait;
 
-    protected function getService(ContainerBuilder $containerBuilder)
+    public function process(ContainerBuilder $container)
     {
-        return $containerBuilder->getDefinition('doppy_pdf_generator.file_locator');
-    }
+        $fileLocatorDefinition = $container->getDefinition('doppy_pdf_generator.file_locator');
+        $fileLocators = $this->findAndSortTaggedServices('doppy_pdf_generator.file_locator', $container);
 
-    protected function getTaggedServices(ContainerBuilder $containerBuilder)
-    {
-        return $containerBuilder->findTaggedServiceIds('doppy_pdf_generator.file_locator');
+        foreach ($fileLocators as $fileLocator) {
+            $fileLocatorDefinition->addMethodCall(
+                'addLocator',
+                [$fileLocator]
+            );
+        }
     }
 }