+++ Exception/ExceptionInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Doppy\PdfGeneratorBundle\Exception;
+
+interface ExceptionInterface {
+
+}