+++ Exception/TempFileGenerationException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Doppy\PdfGeneratorBundle\Exception;
+
+class TempFileGenerationException extends \RuntimeException implements ExceptionInterface
+{
+
+}